Added Articles/Trash
Removed Trash from the filter area so as to improve the UX and reduce confusion as to what happens when you click in the Filter area.
All actions in that area must keap the user on the same page and not work as navigation but only as a filter for the current page.

Signed-off-by: nicohaitas <nicohaitas@gmail.com>